// src/api/generated/model/get-my-attendances-params.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */
export type GetMyAttendancesParams = {
/**
 * Page number (0-indexed)
 * @minimum 0
 */
pageNo?: number;
/**
 * Number of items per page (1-100)
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * Filter by start date
 */
startDate?: string;
/**
 * Filter by end date
 */
endDate?: string;
};